+++ tests/Expressions/jsfmt.spec.js
@@ -38,6 +38,14 @@ describe("Expressions", () => {
         });
         await expect(actual).toMatchFileSnapshot(snapshotFile);
     });
+
+    it("should handle negative number without space", async () => {
+        const { actual, snapshotFile } = await run_spec(import.meta.url, {
+            source: "negative_number.twig"
+        });
+        await expect(actual).toMatchFileSnapshot(snapshotFile);
+    });
+
     it("should handle mapping expressions", async () => {
         const { actual, snapshotFile } = await run_spec(import.meta.url, {
             source: "mappingExpression.twig"